+++ internal/batches/executor/coordinator.go
@@ -14,7 +14,7 @@ import (
 
 type taskExecutor interface {
 	Start(context.Context, []*Task, TaskExecutionUI)
-	Wait(context.Context) ([]taskResult, error)
+	Wait() ([]taskResult, error)
 }
 
 // Coordinator coordinates the execution of Tasks. It makes use of an executor,
@@ -182,7 +182,7 @@ func (c *Coordinator) ExecuteAndBuildSpecs(ctx context.Context, batchSpec *batch
 
 	// Run executor.
 	c.exec.Start(ctx, tasks, ui)
-	results, errs := c.exec.Wait(ctx)
+	results, errs := c.exec.Wait()
 
 	// Write all step cache results to the cache.
 	for _, res := range results {